+++ Quake/Notes005/src/DIYQuake/DIYQuake.cpp
@@ -19,13 +19,13 @@ int main(int argc, char* argv[])
    host.Init(&parameters, &system);
 
    uint64_t OldTime, NewTime;
-   double dTime;
+   float dTime;
    OldTime = system.GetTime();
 
    while (system.GetIsRunning())
    {
       NewTime = system.GetTime();
-      dTime = (double)(NewTime - OldTime) / system.GetPerformanceFrequency();
+      dTime = (float)(NewTime - OldTime) / system.GetPerformanceFrequency();
       host.Frame(dTime);
       OldTime = NewTime;
    }
